[Flat Tourbook View] Partly fixed the reselection after tours are modified, e.g. tags are added or removed.
It is partly fixed because the reselection works but when opening the context menu on the "selected" tours, then mostly the last selected tour gets selected and others are deselected.

I debugged this issue several hours but could not finally fix it.

This fixed a part in this issue  https://github.com/mytourbook/mytourbook/issues/954